Move composition header into new component
